Refactor the write_dataset function
